fix: adjust mla and sla folders and the mean team stamina sla and mla analysis
