--- multiple_log_analysis/__init__.py
@@ -1,7 +1,0 @@
-from .attack.kicks_radial import kicks_radial
-from .ball_possession.ballPossession import BallPossession
-from .io.reader import Reader
-from .utility.finders import find_unique_event_count
-from .utility.finders import find_unique_event_ocurrences
-from .utility.slicers import CustomSlicer
-from .utility.slicers import PlaymodeSlicer
